backend/pyproject.toml
[project]
name = "cs"
version = "0.1.0"
description = "Backend for Coonspect project"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.2",
    "dishka[pytest]>=1.8.0",
    "fastapi[standard]>=0.129.0",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.0",
    "redis>=7.1.1",
    "taskiq-fastapi>=0.4.0",
    "taskiq-redis>=1.2.2",
    "taskiq>=0.12.1",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"] 
flake8-bugbear.extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.14" 
strict = true
ignore_missing_imports = true
exclude = ["tests/"]

explicit_package_bases = true
namespace_packages = true
mypy_path = "." 

[[tool.mypy.overrides]]
module = "motor.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session" 
python_files = ["test_*.py"]
addopts = "--strict-markers --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "app/main.py"]

[tool.coverage.report]
show_missing = true
fail_under = 85  
